SW-149 add Jacoco plugin for code coverage & remove duplicated code
